requirements.txt
bitarray==3.3.1
cryptography==44.0.2
pillow==11.1.0
PySide6==6.9.0